requirements.txt
flask
opencv-python
Pillow
SpeechRecognition
paho-mqtt

django
pandas
matplotlib
seaborn
plotly

fastapi
uvicorn
beautifulsoup4
requests
numpy
# pysqlite3 or brew install sqlite3

# tkinter or brew install python-tk
requests
pygame
pyserial
pyserial-virtual-ports